Add first set of busting functionality
This is the first iteration of my dir busting tool which I'm using
as a learning exercise for Go. It just does top-level folders at
this point but that's good enough for now.

I'm a Go noob, so this stuff can probably be improved a bit, esp
around the concurrency bits.