Only force-append slash to URL in DIR mode
Options changes had code that forced the appending of a forward slash to
the URL prior to continuing. However, this was done for both DIR and DNS
modes. In the case of DNS mode, this would cause a failure because
slashes in DNS requests don't make sense.

This commit modifies the behaviour to include the appending of the slash
only in DIR mode.