Hide progress when Quiet or NoProgress flag set
Added NoProgress as a flag here so that the progress can be hidden from
view.

Also modified the error message for -w so that it includes the reference
to `-w -` in case people forget to set it for stdin.

Added a percentage to the progress output.

Bumped the version number too.