Force mode check to use lowercase
The app crashes if you use "DNS" instead of "dns". This is becauase
we're using `ToLower` in one part of the code and not in another. Once
we're doing with the refactoring of command line options, we're going to
rejig the project structure a bit and make sure that options parsing is
done properly just in the one spot so that we don't have to litter the
code with ToLower to keep things consistent.