Correctly render output in dir mode
When wordlists contain stuff that looks like format strings (such as as
`%7ebob`) the output of the results to screen would contain the string
`%!e(MISSING)`. This was because the entity was passing the word
directly to an Fprintf call as the format string. This PR fixes this so
that the output of the word is rendered correctly.